Seperate i2c_base.py
