+++ ros/riberry_startup/node_scripts/pressure_control_mode.py
@@ -9,13 +9,12 @@ from kxr_controller.kxr_interface import KXRROSRobotInterface
 from kxr_controller.msg import PressureControl
 from skrobot.model import RobotModel
 
-from i2c_for_esp32 import WirePacker
-from filelock import FileLock
-from filelock import Timeout
+from riberry.i2c_base import I2CBase
 
 
-class PressureControlMode(object):
-    def __init__(self, i2c_addr):
+class PressureControlMode(I2CBase):
+    def __init__(self, i2c_addr, lock_path='/tmp/i2c_pressure_control_mode.lock'):
+        super().__init__(i2c_addr)
         # Create robot model to control pressure
         robot_model = RobotModel()
         namespace = ""
@@ -32,7 +31,7 @@ class PressureControlMode(object):
         rospy.Subscriber(
             "/atom_s3_mode", String,
             callback=self.mode_cb, queue_size=1)
-        
+
         # Pressure control
         self.pressure_control_state = {}
         rospy.Subscriber(
@@ -48,45 +47,6 @@ class PressureControlMode(object):
                              callback_args=idx)
             self.pressures[idx] = None
         rospy.Timer(rospy.Duration(0.1), self.timer_callback)
-
-        # I2C
-        self.i2c_addr = i2c_addr
-        self.device_type = identify_device()
-        if self.device_type == 'Raspberry Pi':
-            import board
-            import busio
-            self.i2c = busio.I2C(board.SCL, board.SDA)
-            bus_number = 1
-        elif self.device_type == 'Radxa Zero':
-            import board
-            import busio
-            self.i2c = busio.I2C(board.SCL1, board.SDA1)
-            bus_number = 3
-        elif self.device_type == 'Khadas VIM4':
-            self.i2c = i2c()
-            bus_number = None
-        else:
-            raise ValueError('Unknown device {}'.format(
-                self.device_type))
-        self.lock = FileLock(lock_path, timeout=10)
-
-    def i2c_write(self, packet):
-        try:
-            self.lock.acquire()
-        except Timeout as e:
-            print(e)
-            return
-        try:
-            self.i2c.writeto(self.i2c_addr, packet)
-        except OSError as e:
-            print(e)
-        except TimeoutError as e:
-            print('I2C Write error {}'.format(e))
-        try:
-            self.lock.release()
-        except Timeout as e:
-            print(e)
-            return
 
     def button_cb(self, msg):
         """
@@ -104,7 +64,7 @@ class PressureControlMode(object):
         Check AtomS3 mode.
         """
         self.mode = msg.data
-            
+
     def pressure_control_cb(self, msg):
         self.pressure_control_state[f'{msg.board_idx}'] = msg
 
@@ -144,35 +104,7 @@ class PressureControlMode(object):
                 sent_str += '{}: {:.3f}\n'.format(idx, value)
         sent_str += '\nSingle Click:\nVacuum on off'
         # Send message on AtomS3 LCD
-        packer = WirePacker(buffer_size=len(sent_str) + 8)
-        for s in sent_str:
-            packer.write(ord(s))
-        packer.end()
-        if packer.available():
-            self.i2c_write(packer.buffer[:packer.available()])
-
-lock_path = '/tmp/i2c-1.lock'
-
-def identify_device():
-    try:
-        with open('/proc/cpuinfo', 'r') as f:
-            cpuinfo = f.read()
-
-        if 'Raspberry Pi' in cpuinfo:
-            return 'Raspberry Pi'
-
-        with open('/proc/device-tree/model', 'r') as f:
-            model = f.read().strip()
-
-        # remove null character
-        model = model.replace('\x00', '')
-
-        if 'Radxa' in model or 'ROCK Pi' in model or model in 'Khadas VIM4':
-            return model
-
-        return 'Unknown Device'
-    except FileNotFoundError:
-        return 'Unknown Device'
+        self.send_string(sent_str)
 
 
 if __name__ == '__main__':
